Revert "New more info light (#14453)"
This reverts commit 7c4a421e74b84c53e1487f3fdfc172f50dd4f65e.
