+++ package.json
@@ -71,7 +71,6 @@
     "@material/mwc-textfield": "^0.27.0",
     "@material/mwc-top-app-bar-fixed": "^0.27.0",
     "@material/top-app-bar": "=14.0.0-canary.53b3cad2f.0",
-    "@material/web": "^1.0.0-pre.1",
     "@mdi/js": "7.1.96",
     "@mdi/svg": "7.1.96",
     "@polymer/app-layout": "^3.1.0",
